animation system BITCH

+++ pge-core/include/PGE/core/pgeCore.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "PGE/utils/pgeLogger.h"
+#include "PGE/utils/pgeTypes.h"
 
 #include <SDL2/SDL.h>
 
@@ -8,6 +9,6 @@ namespace pge
 {
     namespace core
     {
-        int Init();
+        i8 Init();
     }
 }
